Keep API request functions in main Service file
